Move `EventCatcher` to `tests.utils` for use in unit and functional tests
