+++ tests/unit/events/test_logging.py
@@ -8,7 +8,7 @@ from dbt.flags import get_flags, set_from_args
 from dbt_common.events.base_types import BaseEvent
 from dbt_common.events.event_manager_client import get_event_manager
 from dbt_common.events.logger import LoggerConfig
-from tests.functional.utils import EventCatcher
+from tests.utils import EventCatcher
 
 
 class TestSetupEventLogger:
